Convert deploy to use logging library
This converts the deploy.js script to use the logging library.
Additionally, the flags library was updated to have a Counter type,
which was utilized to add a verbose flag to the deploy script. This
allows the script to handle short opts like: `-vvv` to increase the
verbosity of the logging.