Move library scripts to `lib/`
This moves library scripts to the `/lib/` subdirectory. This should make
it easier to discover new libraries as they are added for scripts that
need to do so.

This also updates the `tsconfig.json` file to properly include
`./exploit` and `./lib` files, and turns on type checking for the
remaining scripts.